+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
-## [Unreleased]
+
+## [1.0.1] - 2023-12-11
+### Changed
+- Lock thumbnail image to 1.0.0
 
 ## [1.0.0] - 2023-12-11
 ### Added
@@ -25,7 +28,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - Initiated project
 
-[Unreleased]: https://github.com/rarimo/collabland-rarimo-poh-action/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/rarimo/collabland-rarimo-poh-action/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/rarimo/collabland-rarimo-poh-action/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/rarimo/collabland-rarimo-poh-action/compare/0.1.0...1.0.0
 [0.1.0]: https://github.com/rarimo/collabland-rarimo-poh-action/releases/tag/0.1.0
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/